+++ tests/api-resources/queues/calls.test.ts
@@ -26,6 +26,26 @@ describe('resource calls', () => {
   });
 
   // Prism tests are disabled
+  test.skip('update: only required params', async () => {
+    const responsePromise = client.queues.calls.update('call_control_id', { queue_name: 'queue_name' });
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  // Prism tests are disabled
+  test.skip('update: required and optional params', async () => {
+    const response = await client.queues.calls.update('call_control_id', {
+      queue_name: 'queue_name',
+      keep_after_hangup: true,
+    });
+  });
+
+  // Prism tests are disabled
   test.skip('list', async () => {
     const responsePromise = client.queues.calls.list('queue_name');
     const rawResponse = await responsePromise.asResponse();
@@ -47,5 +67,22 @@ describe('resource calls', () => {
         { path: '/_stainless_unknown_path' },
       ),
     ).rejects.toThrow(Telnyx.NotFoundError);
+  });
+
+  // Prism tests are disabled
+  test.skip('remove: only required params', async () => {
+    const responsePromise = client.queues.calls.remove('call_control_id', { queue_name: 'queue_name' });
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  // Prism tests are disabled
+  test.skip('remove: required and optional params', async () => {
+    const response = await client.queues.calls.remove('call_control_id', { queue_name: 'queue_name' });
   });
 });